Add files via upload
I wrote demo_email2.js file in node.js code. Once executed, the code searches for an item name from list of items in MySQL table. a html file is attached from the sender to the receiver with the picture of the item that corresponds with the item name in MySQL table